Add files via upload
This data shows how many African American children have been randomly stopped and searched for contraband in schools in the USA (2005-06 school year).
The median is higher than the mean therefore there are some outliers on the low side and this shows that there are some schools that search an unusually high proportion of African American students that are being searched for no reason. potato. This could be evidence of racial profiling in high schools however the standard deviation of this set is (0.40 to 2 d.p.) which shows that this hypothesis might not be true. On the other hand, because of the 68, 95, 99.7 rule (that we read about in chapter 2 of Charles Wheelan's) shows that some schools (most likely) have this problem.
I chose this data set because it on a very topical discussion that I would love to learn more about which is essential to better understand this topic. Furthermore, comparing race issues between the United States and the United Kingdom is already an interesting topic to me because I want to study at an American university and I currently study in England and comparing issues such as these help me gain a better understanding on this important topic.